renaming PHP_Repl to PHPRepl
